01/01/2025 AS 20:52 ALTERAÇÃO DE ÍCONES, ICONOGRAFIA QUASE PRONTA

+++ MODEL/ItensVendaMODEL.cs
@@ -9,12 +9,12 @@ namespace Money
     internal class ItensVendaMODEL
     {
         private int id_itensvenda, id_produto, qtd_produto, id_venda, num_parcela;
-        private double valor_produto;
+        private decimal valor_produto;
 
         public int Id_itensvenda { get => Id_itensvenda1; set => Id_itensvenda1 = value; }
         public int Id_produto { get => Id_produto1; set => Id_produto1 = value; }
         public int Qtd_produto { get => Qtd_produto1; set => Qtd_produto1 = value; }
-        public double Valor_produto { get => valor_produto; set => valor_produto = value; }
+        public decimal Valor_produto { get => valor_produto; set => valor_produto = value; }
         public int Id_produto1 { get => Id_produto2; set => Id_produto2 = value; }
         
         public int Id_venda { get => Id_venda1; set => Id_venda1 = value; }
